requirements.txt
# Django Core
Django==5.2.12
djangorestframework==3.14.0
django-environ==0.10.0
python-decouple==3.8
django-cors-headers==4.3.1

# JWT Authentication
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# PDF Generation & Image Processing
reportlab==4.0.7
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2

# WhatsApp Integration
twilio==8.10.0

# Development (opcional)
django-extensions==3.2.3
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Production (opcional)
gunicorn==21.2.0
whitenoise==6.6.0
